Move scroll behaviour into MainHeader
