+++ front/app/containers/App/index.tsx
@@ -29,7 +29,6 @@ import Navigate from 'utils/cl-router/Navigate';
 import Authentication from 'containers/Authentication';
 import MainHeader from 'containers/MainHeader';
 import Meta from './Meta';
-import MobileScrollHeader from 'containers/MainHeader/Components/MobileScrollHeader';
 const UserDeletedModal = lazy(() => import('./UserDeletedModal'));
 const PlatformFooter = lazy(() => import('containers/PlatformFooter'));
 
@@ -341,7 +340,6 @@ const App = ({ children }: Props) => {
               {showFrontOfficeNavbar && (
                 <ErrorBoundary>
                   <MainHeader />
-                  <MobileScrollHeader />
                 </ErrorBoundary>
               )}
               <Box
